QOL Additions and English improvements
Before starting work on the expedition system part 2, I added new QOL things to the game. which are: UI to show the tree life (health bar), a flexible buy system, a UI for show our coins. I thought it was interesting to add a flexible buy system before I start to work on the expedition system, which will use so many coins and items...

Sorry for the English mistakes in my previous posts... Im practicing it slowly :)

+++ LumberjackGame/Assets/Scripts/PlayerSystems/InventorySystem/InventoryItem.cs
@@ -1,28 +1,25 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 
 [Serializable]
 public class InventoryItem
 {
     public int ID;
     public Item itemData;
-    public int stackSize;
+    public float stackSize;
 
-    public InventoryItem(int _id, Item item, int amount)
+    public InventoryItem(int _id, Item item, float amount)
     {
         ID = _id;
         itemData = item;
         AddToStack(amount);
     }
 
-    public void AddToStack(int add)
+    public void AddToStack(float add)
     {
         stackSize += add;
     }
 
-    public void RemoveFromStack(int remove) 
+    public void RemoveFromStack(float remove) 
     {
         stackSize -= remove;
     }
